Add: [화면] #194 Vue > Nuxt 이동 ( Leaf, Member 부분 )
