--- frontend-2/pages/app/leaf/LeafBookEdit.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-</script>
-
-<template>
-</template>
-
-<style scoped>
-
-</style>
