--- frontend-2/pages/app/leaf/LeafBookRegister.vue
@@ -1,5 +1,0 @@
-<script setup lang="ts"></script>
-
-<template></template>
-
-<style scoped></style>
