--- frontend-2/pages/app/leaf/LeafDetail.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  Leaf Detail
-</template>
-
-<style scoped>
-
-</style>
