--- frontend-2/pages/app/leaf/LeafEtcEdit.vue
@@ -1,11 +1,0 @@
-<script setup>
-
-</script>
-
-<template>
-
-</template>
-
-<style scoped>
-
-</style>
