--- frontend-2/pages/app/leaf/LeafEtcRegister.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  Leaf Register
-</template>
-
-<style scoped>
-
-</style>
