--- frontend-2/pages/app/leaf/LeafIndex.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <RouterView />
-</template>
-
-<style scoped>
-
-</style>
