--- frontend-2/pages/app/member/MemberIndex.vue
@@ -1,12 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <RouterView />
-</template>
-
-<style scoped>
-
-
-</style>
