--- frontend-2/pages/app/trees/TreeList.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  Tree List
-</template>
-
-<style scoped>
-
-</style>
