--- frontend-2/pages/app/trees/TreeSearch.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-</script>
-
-<template>
-  Tree Search
-</template>
-
-<style scoped>
-
-</style>
