--- frontend-2/pages/app/trees/TreeSearchResult.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  Tree Search Result
-</template>
-
-<style scoped>
-
-</style>
