--- frontend-2/pages/app/trees/[id]/treeMemoirRegisterIndex.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-</script>
-
-<template>
-  Tree Memoir Register Index
-</template>
-
-<style scoped>
-
-</style>
