--- frontend-2/pages/app/trees/index.vue
@@ -1,13 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <main>
-    <h1>Trees</h1>
-  </main>
-</template>
-
-<style scoped>
-
-</style>
